python/requirements.txt
beautifulsoup4==4.12.3
pypresence==4.6.1
requests==2.32.3
obsws-python==1.7.1
pillow==12.0.0
imagehash==4.3.1
numpy==2.3.5
scipy==1.16.3
discord-webhook==1.4.1
keyboard==0.13.5
pyautogui==0.9.54
pyinstaller==6.17.0
easyocr=1.7.2

# PySimpleGUI (use static URL below)
# Older version of PySimpleGUI since 5.x has been discontinued
PySimpleGUI @ https://github.com/egm3387/PySimpleGUI-4.60.5/raw/refs/heads/main/PySimpleGUI-4.60.5.tar.gz